requirements.txt
gradio
transformers
accelerate
huggingface_hub
PyYAML